Move label_and_render_plot to plot.py
